# Upgrade to Esprima 3.0.0-beta-7