--- a/Jint/Native/Function/ScriptFunctionInstance.cs
+++ b/Jint/Native/Function/ScriptFunctionInstance.cs
@@ -23,7 +23,7 @@ namespace Jint.Native.Function
             ObjectInstance? proto = null)
             : this(
                 engine,
-                new JintFunctionDefinition(engine, functionDeclaration),
+                new JintFunctionDefinition(functionDeclaration),
                 scope,
                 strict ? FunctionThisMode.Strict : FunctionThisMode.Global,
                 proto)
@@ -39,7 +39,7 @@ namespace Jint.Native.Function
             : base(engine, engine.Realm, function, scope, thisMode)
         {
             _prototype = proto ?? _engine.Realm.Intrinsics.Function.PrototypeObject;
-            _length = new LazyPropertyDescriptor(null, _ => JsNumber.Create(function.Initialize(this).Length), PropertyFlag.Configurable);
+            _length = new LazyPropertyDescriptor(null, _ => JsNumber.Create(function.Initialize().Length), PropertyFlag.Configurable);
 
             if (!function.Strict
                 && !engine._isStrict