--- a/Jint/Runtime/Interpreter/Expressions/JintFunctionExpression.cs
+++ b/Jint/Runtime/Interpreter/Expressions/JintFunctionExpression.cs
@@ -8,9 +8,9 @@ namespace Jint.Runtime.Interpreter.Expressions
     {
         private readonly JintFunctionDefinition _function;
 
-        public JintFunctionExpression(Engine engine, FunctionExpression function) : base(function)
+        public JintFunctionExpression(FunctionExpression function) : base(function)
         {
-            _function = new JintFunctionDefinition(engine, function);
+            _function = new JintFunctionDefinition(function);
         }
 
         protected override ExpressionResult EvaluateInternal(EvaluationContext context)