--- a/Jint/Runtime/Interpreter/Expressions/JintIdentifierExpression.cs
+++ b/Jint/Runtime/Interpreter/Expressions/JintIdentifierExpression.cs
@@ -1,3 +1,5 @@
+using System.Diagnostics.CodeAnalysis;
+using System.Runtime.CompilerServices;
 using Esprima.Ast;
 using Jint.Native;
 using Jint.Native.Argument;
@@ -7,31 +9,27 @@ namespace Jint.Runtime.Interpreter.Expressions
 {
     internal sealed class JintIdentifierExpression : JintExpression
     {
-        internal readonly EnvironmentRecord.BindingName _expressionName;
-        private readonly JsValue? _calculatedValue;
-
         public JintIdentifierExpression(Identifier expression) : base(expression)
         {
-            _expressionName = new EnvironmentRecord.BindingName(expression.Name);
-            if (expression.Name == "undefined")
-            {
-                _calculatedValue = JsValue.Undefined;
-            }
         }
 
-        public bool HasEvalOrArguments
-            => _expressionName.Key == KnownKeys.Eval || _expressionName.Key == KnownKeys.Arguments;
+        internal EnvironmentRecord.BindingName Identifier
+        {
+            get => (EnvironmentRecord.BindingName) (_expression.AssociatedData ??= new EnvironmentRecord.BindingName(((Identifier) _expression).Name));
+        }
+
+        public bool HasEvalOrArguments => Identifier.HasEvalOrArguments;
 
         protected override ExpressionResult EvaluateInternal(EvaluationContext context)
         {
             var engine = context.Engine;
             var env = engine.ExecutionContext.LexicalEnvironment;
             var strict = StrictModeScope.IsStrictModeCode;
-            var identifierEnvironment = JintEnvironment.TryGetIdentifierEnvironmentWithBinding(env, _expressionName, out var temp)
+            var identifierEnvironment = JintEnvironment.TryGetIdentifierEnvironmentWithBinding(env, Identifier, out var temp)
                 ? temp
                 : JsValue.Undefined;
 
-            return NormalCompletion(engine._referencePool.Rent(identifierEnvironment, _expressionName.StringValue, strict, thisValue: null));
+            return NormalCompletion(engine._referencePool.Rent(identifierEnvironment, Identifier.StringValue, strict, thisValue: null));
         }
 
         public override Completion GetValue(EvaluationContext context)
@@ -39,9 +37,9 @@ namespace Jint.Runtime.Interpreter.Expressions
             // need to notify correct node when taking shortcut
             context.LastSyntaxElement = _expression;
 
-            if (_calculatedValue is not null)
+            if (Identifier.CalculatedValue is not null)
             {
-                return new(CompletionType.Normal, _calculatedValue, _expression);
+                return new(CompletionType.Normal, Identifier.CalculatedValue, _expression);
             }
 
             var strict = StrictModeScope.IsStrictModeCode;
@@ -50,19 +48,19 @@ namespace Jint.Runtime.Interpreter.Expressions
 
             if (JintEnvironment.TryGetIdentifierEnvironmentWithBindingValue(
                 env,
-                _expressionName,
+                Identifier,
                 strict,
                 out _,
                 out var value))
             {
                 if (value is null)
                 {
-                    ExceptionHelper.ThrowReferenceError(engine.Realm, _expressionName.Key.Name + " has not been initialized");
+                    ThrowNotInitialized(engine);
                 }
             }
             else
             {
-                var reference = engine._referencePool.Rent(JsValue.Undefined, _expressionName.StringValue, strict, thisValue: null);
+                var reference = engine._referencePool.Rent(JsValue.Undefined, Identifier.StringValue, strict, thisValue: null);
                 value = engine.GetValue(reference, true);
             }
 
@@ -73,6 +71,13 @@ namespace Jint.Runtime.Interpreter.Expressions
             }
 
             return new(CompletionType.Normal, value, _expression);
+        }
+
+        [DoesNotReturn]
+        [MethodImpl(MethodImplOptions.NoInlining)]
+        private void ThrowNotInitialized(Engine engine)
+        {
+            ExceptionHelper.ThrowReferenceError(engine.Realm, Identifier.Key.Name + " has not been initialized");
         }
     }
 }