… Jint/Runtime/Interpreter/Expressions/JintObjectExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintUpdateExpression.cs
… Jint/Runtime/Interpreter/JintFunctionDefinition.cs
… Jint/Runtime/Interpreter/JintStatementList.cs
… Jint/Runtime/KnownKeys.cs
… Jint/Runtime/Modules/SourceTextModuleRecord.cs