* add hooks to store static data to nodes
* make JintFunctionDefinition less bound to engine
* code changes to improve compatibility