@lahma 2022-09-09
Still a bit of work in progress, but shows promise with existing (prepared) script.

## Jint.Benchmark.EvaluationBenchmark

| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |Execute|28.76 μs|0.180 μs|33.97 KB|
| **New** |	| **29.51 μs (+3%)** | **0.119 μs** | **34.48 KB (+2%)** |
| Old |Execute_ParsedScript|12.94 μs|0.048 μs|26.13 KB|
| **New** |	| **12.22 μs (-6%)** | **0.224 μs** | **24.98 KB (-4%)** |